Moved node-comments classes to their own namespace
