+++ src/TypedTree.Generator.Core/Mapping/Context.cs
@@ -2,6 +2,7 @@
 using System.Text.RegularExpressions;
 using Microsoft.Extensions.Logging;
 
+using TypedTree.Generator.Core.Mapping.NodeComments;
 using TypedTree.Generator.Core.Utilities;
 
 namespace TypedTree.Generator.Core.Mapping
