+++ src/TypedTree.Generator.Tests/Mapping/NodeComments/ManualCommentProviderTests.cs
@@ -1,9 +1,9 @@
 using System.Collections.Generic;
 using Xunit;
 
-using TypedTree.Generator.Core.Mapping;
+using TypedTree.Generator.Core.Mapping.NodeComments;
 
-namespace TypedTree.Generator.Tests.Mapping
+namespace TypedTree.Generator.Tests.Mapping.NodeComments
 {
     public sealed class ManualCommentProviderTests
     {
